requirements.txt
Flask
gunicorn
mysql-connector-python
werkzeug
flask_redmail
flask-mail
# smtplib
# flask-cors
